+++ py/desispec/io/frame.py
@@ -15,10 +15,9 @@ from desiutil.depend import add_dependencies
 from desiutil.io import encode_table
 
 from ..frame import Frame
-from .meta import findfile
+from .meta import findfile, get_nights, get_exposures
 from .util import fitsheader, native_endian, makepath
 from desiutil.log import get_logger
-
 
 def write_frame(outfile, frame, header=None, fibermap=None, units=None):
     """Write a frame fits file and returns path to file written.
@@ -91,18 +90,21 @@ def write_frame(outfile, frame, header=None, fibermap=None, units=None):
 
     return outfile
 
-def read_meta_frame(filename):
-    """ 
+
+def read_meta_frame(filename, extname=0):
+    """ Load the meta information of a Frame
     Args:
-        filename: path to a file, or (night, expid, camera) tuple where
+        filename: path to a file
+        extname: int, optional;  Extension for grabbing header info
 
     Returns:
         meta: dict or astropy.fits.header
 
     """
     fx = fits.open(filename, uint=True, memmap=False)
-    hdr = fx[0].header
+    hdr = fx[extname].header
     return hdr
+
 
 def read_frame(filename, nspec=None):
     """Reads a frame fits file and returns its data.
@@ -169,3 +171,32 @@ def read_frame(filename, nspec=None):
         log.error("Frame did not pass simple vetting test. diagnosis={:d}".format(diagnosis))
     # Return
     return frame
+
+
+def search_for_framefile(frame_file):
+    """ Search for an input frame_file in the desispec redux hierarchy
+    Args:
+        frame_file:  str
+
+    Returns:
+        mfile: str,  full path to frame_file if found else raise error
+
+    """
+    log=get_logger()
+    # Parse frame file name
+    ifile = frame_file.split('/')[-1]
+    splits = ifile.split('-')
+    root = splits[0]
+    camera = splits[1]
+    fexposure = int(splits[2].split('.')[0])
+
+    # Loop on nights
+    nights = get_nights()
+    for night in nights:
+        for exposure in get_exposures(night):
+            if exposure == fexposure:
+                mfile = findfile(root, camera=camera, night=night, expid=exposure)
+                if os.path.isfile(mfile):
+                    return mfile
+                else:
+                    log.error("Expected file {:s} not found..".format(mfile))